cleanup(test): Clean up test files   * Rebuilt the project

+++ configs/buildHook.js
@@ -3,6 +3,7 @@ import { exec } from 'child_process'
 import path from 'path'
 
 const cmdExec = promisify(exec)
+const { NODE_ENV } = process.env
 
 /**
  * 
@@ -15,12 +16,16 @@ export default function buildHook(){
   return {
     name: 'buildHook',
     buildEnd: async () => {
+      if(NODE_ENV !== 'development') return
+
       const buildDir = path.join(__dirname, '../build/cjs')
       const docsDir = path.join(__dirname, '../docs/parkin')
     
+      console.log(`Removing old docs/parkin folder...`)
       // Remove the old docks build directory
       await cmdExec(`rm -rf ${docsDir}`)
 
+      console.log(`Copying build/cjs into docs/parkin folder...`)
       // Copy over the new build directory
       await cmdExec(`cp -R ${buildDir} ${docsDir}`)
